requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
werkzeug==3.0.1
itsdangerous==2.1.2
pandas==2.2.3
reportlab==4.0.7
requests==2.31.0
PyMySQL==1.1.0
psycopg==3.2.1
cryptography==41.0.7
python-dotenv==1.0.1
mysql-connector-python==9.0.0
numpy==2.1.2
groq==0.8.0
httpx==0.27.2

